Move generic source files to packages; change to snake_case
